Updates for flexible filament.
The case doesn't have the same top/bottom cutouts anymore, because it doesn't
have to be made to be more flexible. So therefore the top, bottom and right
cutouts are now all bridged. It is best to print with supports so that those
four bridges come out perfect.

This is intended to work with a flexible filament.

At 0.8mm walls it can work with Nylon (tested with Taulman Bridge) but
sometimes tears between layers and you have to use a brim.

At 1.2mm it works with some noname TPE I have that's probably similar to
semiflex. At 0.8mm the flexible filament comes out really bad and actually too
flexible, at 1.2mm it is the most perfect thing that ever came out of my 3D
printer. Closest to something that could have been bought in a store ;)

Would love to try this with different size nozzles so I can get to 1.0mm walls
and see how that turns out.